backend/src/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.8
django-cors-headers==3.11.0
djangorestframework==3.13.1
#psycopg2==2.9.3
psycopg2-binary==2.9.5
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
pyjwt-2.6.0
pycparser-2.21